move smoother to tikv_util
